Bug 1215664: Merged separated model files into one. r=mcote
