+++ pulseguardian/model/base.py
@@ -23,3 +23,6 @@ Base.query = db_session.query_property()
 
 def init_db():
     Base.metadata.create_all(bind=engine)
+
+def drop_db():
+    Base.metadata.drop_all(bind=engine)
