--- pulseguardian/model/queue.py
@@ -1,24 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
-
-from pulseguardian.model.base import Base
-
-
-class Queue(Base):
-    __tablename__ = 'queues'
-
-    name = Column(String(255), primary_key=True)
-    owner_id = Column(Integer, ForeignKey('pulse_users.id'), nullable=True)
-    size = Column(Integer)
-
-    warned = Column(Boolean)
-
-    durable = Column(Boolean, nullable=False, default=False)
-
-    def __repr__(self):
-        return "<Queue(name='{0}', owner='{1}')>".format(self.name, self.owner)
-
-    __str__ = __repr__
